+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 
 - [core] feature: display the `job class` property when inspecting a `job` object.  
 - [warning] feature: staff notifications now include only active warnings.  
-- [core] feature: improve the `/fuji inspect argument-types` command, to distinguish the `vanilla argument types` and `non-vanilla argument types` more precise.
+- [core] docs: improve the `/fuji inspect argument-types` command, to distinguish the `vanilla argument types` and `non-vanilla argument types` more precise.
+- [core] docs: add the document for `argument type adapters` in the inspector GUI.